translator/requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cloudmersive-convert-api-client==3.2.5
filelock==3.12.0
fire==0.5.0
Flask==2.3.1
fsspec==2023.4.0
fugashi==1.2.1
google==3.0.0
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
googletrans==3.1.0a0
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.14.1
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.6.0
itsdangerous==2.1.2
jaconv==0.3.4
Jinja2==3.1.2
loguru==0.7.0
manga-ocr==0.1.8
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyperclip==1.8.2
pytesseract==0.3.10
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
rfc3986==1.5.0
rsa==4.9
sentencepiece==0.1.98
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sympy==1.11.1
termcolor==2.3.0
tesseract==0.1.3
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
unidic-lite==1.0.8
urllib3==1.26.15
Werkzeug==2.3.1
zipp==3.15.0
